pyproject.toml
[project]
name = "curio"
version = "0.1.0"
description = "Curio media library"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.5",
    "django-allauth>=65.16.1",
    "django-vite>=3.1.0",
    "mutagen>=1.47.0",
    "psycopg>=3.3.4",
]

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "curio.settings"
python_files = ["tests.py", "test_*.py"]

[tool.djlint]
profile = "django"
indent = 2

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
    "pytest-django>=4.12.0",
]
